requirements.txt
annotated-types==0.7.0
anyio==4.6.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
boto3==1.35.20
botocore==1.35.20
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==5.1
django-allauth==65.0.1
django-browser-reload==1.15.0
django-parler==2.3
django-storages==1.14.4
django-tailwind==3.8.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
jmespath==1.0.1
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
replicate==0.33.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
socksio==1.0.0
sqlparse==0.5.1
stripe==10.12.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
urllib3==2.2.2
